Part 2-Tensorflow Balanced Data/requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/dev.in
#
absl-py==0.15.0
    # via
    #   tensorboard
    #   tensorflow-gpu
addict==2.4.0
    # via -r requirements/dev.in
albumentations==1.1.0
    # via -r requirements/dev.in
astunparse==1.6.3
    # via tensorflow-gpu
cachetools==4.2.4
    # via google-auth
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via wandb
configparser==5.2.0
    # via wandb
cycler==0.11.0
    # via matplotlib
distlib==0.3.4
    # via virtualenv
docker-pycreds==0.4.0
    # via wandb
efficientnet==1.0.0
    # via segmentation-models
filelock==3.4.2
    # via virtualenv
flatbuffers==1.12
    # via tensorflow-gpu
fonttools==4.28.3
    # via matplotlib
gast==0.4.0
    # via tensorflow-gpu
gitdb==4.0.9
    # via gitpython
gitpython==3.1.24
    # via wandb
google-auth==2.3.3
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow-gpu
grpcio==1.34.1
    # via
    #   tensorboard
    #   tensorflow-gpu
h5py==3.1.0
    # via
    #   keras-applications
    #   tensorflow-gpu
identify==2.4.8
    # via pre-commit
idna==3.3
    # via requests
image-classifiers==1.0.0
    # via segmentation-models
imageio==2.13.3
    # via scikit-image
importlib-metadata==4.8.2
    # via markdown
joblib==1.1.0
    # via scikit-learn
keras-applications==1.0.8
    # via
    #   efficientnet
    #   image-classifiers
    #   segmentation-models
keras-nightly==2.5.0.dev2021032900
    # via tensorflow-gpu
keras-preprocessing==1.1.2
    # via tensorflow-gpu
kiwisolver==1.3.2
    # via matplotlib
markdown==3.3.6
    # via tensorboard
matplotlib==3.5.1
    # via -r requirements/dev.in
networkx==2.6.3
    # via scikit-image
nodeenv==1.6.0
    # via pre-commit
numpy==1.19.5
    # via
    #   albumentations
    #   h5py
    #   imageio
    #   keras-applications
    #   keras-preprocessing
    #   matplotlib
    #   opencv-python
    #   opencv-python-headless
    #   opt-einsum
    #   pandas
    #   pywavelets
    #   qudida
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow-gpu
    #   tifffile
oauthlib==3.1.1
    # via requests-oauthlib
opencv-python==4.5.4.60
    # via -r requirements/dev.in
opencv-python-headless==4.5.4.60
    # via
    #   albumentations
    #   qudida
opt-einsum==3.3.0
    # via tensorflow-gpu
packaging==21.3
    # via
    #   matplotlib
    #   scikit-image
pandas==1.3.5
    # via -r requirements/dev.in
pathtools==0.1.2
    # via wandb
pillow==8.4.0
    # via
    #   imageio
    #   matplotlib
    #   scikit-image
platformdirs==2.4.1
    # via virtualenv
pre-commit==2.17.0
    # via -r requirements/dev.in
promise==2.3
    # via wandb
protobuf==3.19.1
    # via
    #   tensorboard
    #   tensorflow-gpu
    #   wandb
psutil==5.8.0
    # via wandb
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.6
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
    #   wandb
pytz==2021.3
    # via pandas
pywavelets==1.2.0
    # via scikit-image
pyyaml==6.0
    # via
    #   albumentations
    #   pre-commit
    #   wandb
qudida==0.0.4
    # via albumentations
requests==2.26.0
    # via
    #   requests-oauthlib
    #   tensorboard
    #   wandb
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
scikit-image==0.19.0
    # via
    #   albumentations
    #   efficientnet
scikit-learn==1.0.1
    # via
    #   -r requirements/dev.in
    #   qudida
scipy==1.7.3
    # via
    #   albumentations
    #   scikit-image
    #   scikit-learn
segmentation-models==1.0.1
    # via -r requirements/dev.in
sentry-sdk==1.5.1
    # via wandb
shortuuid==1.0.8
    # via wandb
six==1.15.0
    # via
    #   absl-py
    #   astunparse
    #   docker-pycreds
    #   google-auth
    #   google-pasta
    #   grpcio
    #   keras-preprocessing
    #   promise
    #   python-dateutil
    #   tensorflow-gpu
    #   virtualenv
    #   wandb
smmap==5.0.0
    # via gitdb
subprocess32==3.5.4
    # via wandb
tensorboard==2.7.0
    # via tensorflow-gpu
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.0
    # via tensorboard
tensorflow-addons==0.15.0
    # via -r requirements/dev.in
tensorflow-estimator==2.5.0
    # via tensorflow-gpu
tensorflow-gpu==2.5
    # via -r requirements/dev.in
termcolor==1.1.0
    # via
    #   tensorflow-gpu
    #   yaspin
threadpoolctl==3.0.0
    # via scikit-learn
tifffile==2021.11.2
    # via scikit-image
toml==0.10.2
    # via pre-commit
typeguard==2.13.3
    # via tensorflow-addons
typing-extensions==3.7.4.3
    # via
    #   gitpython
    #   qudida
    #   tensorflow-gpu
urllib3==1.26.7
    # via
    #   requests
    #   sentry-sdk
virtualenv==20.13.1
    # via pre-commit
wandb==0.12.9
    # via -r requirements/dev.in
werkzeug==2.0.2
    # via tensorboard
wheel==0.37.0
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow-gpu
wrapt==1.12.1
    # via tensorflow-gpu
yaspin==2.1.0
    # via wandb
zipp==3.6.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
